Updated Jenkinsfile

+++ Jenkinsfile
@@ -5,7 +5,16 @@ pipeline {
         AWS_REGION = 'us-east-1'
     }
 
+    parameters {
+        choice(
+            name: 'ACTION',
+            choices: ['apply', 'destroy'],
+            description: 'Choose Terraform action to perform'
+        )
+    }
+
     stages {
+
         stage('Code Checkout') {
             steps {
                 git branch: 'main', url: 'https://github.com/Roshanx96/demo-terraform.git'
@@ -43,20 +52,19 @@ pipeline {
             }
         }
 
-        stage('Terraform Apply') {
+        stage('Terraform Apply or Destroy') {
             steps {
                 script {
                     withAWS(credentials: 'aws-cred', region: "${AWS_REGION}") {
-                        sh 'terraform apply -auto-approve'
-                    }
-                }
-            }
-        }
-        stage('Terraform Destroy') {
-            steps {
-                script {
-                    withAWS(credentials: 'aws-cred', region: "${AWS_REGION}") {
-                        sh 'terraform destroy -auto-approve'
+                        if (params.ACTION == 'apply') {
+                            echo '🚀 Running Terraform Apply...'
+                            sh 'terraform apply -auto-approve'
+                        } else if (params.ACTION == 'destroy') {
+                            echo '🔥 Running Terraform Destroy...'
+                            sh 'terraform destroy -auto-approve'
+                        } else {
+                            error("❌ Invalid action: ${params.ACTION}")
+                        }
                     }
                 }
             }
@@ -65,10 +73,10 @@ pipeline {
 
     post {
         success {
-            echo '✅ Pipeline executed successfully!'
+            echo "✅ Terraform ${params.ACTION} executed successfully!"
         }
         failure {
-            echo '❌ Pipeline failed. Please check the logs.'
+            echo "❌ Terraform ${params.ACTION} failed. Please check the logs."
         }
     }
 }
